+++ layout.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"image"
+	//	"image"
 
 	"github.com/as/ui/tag"
 	"github.com/as/ui/win"
@@ -11,12 +11,10 @@ import (
 var (
 	winid  = make(map[int]*win.Win)
 	wincol = make(map[int]*Col)
-	xx     Cursor
 	down   uint
-	pt     image.Point
 )
 
-func readmouse(e mouse.Event) mouse.Event {
+func readmouse0(e mouse.Event) mouse.Event {
 	switch e.Direction {
 	case 1:
 		down |= 1 << uint(e.Button)
@@ -24,100 +22,38 @@ func readmouse(e mouse.Event) mouse.Event {
 		down &^= 1 << uint(e.Button)
 	}
 	activate(p(e), g)
-	pt = p(e).Add(act.Loc().Min)
-	e.X -= float32(act.Sp.X)
-	e.Y -= float32(act.Sp.Y)
 	return e
 }
 
 func (g *Grid) dragCol(c *Col, e mouse.Event, mousein <-chan mouse.Event) {
-	xx.srcCol = actCol
-	xx.sweepCol = true
-	xx.src = nil
-	g.detach(g.ID(xx.srcCol))
+	c0 := actCol
+	g.detach(g.ID(c0))
 	g.fill()
-
 	for e = range mousein {
-		e = readmouse(e)
+		e = readmouse0(e)
 		if down == 0 {
 			break
 		}
 	}
-	e.X += float32(act.Sp.X)
-	e.Y += float32(act.Sp.Y)
-	pt.X = int(e.X)
-	pt.Y = int(e.Y)
 	activate(p(e), g)
-
-	xx.sweepCol = false
 	g.fill()
-	g.Attach(xx.srcCol, pt.X)
-	moveMouse(xx.srcCol.Loc().Min)
+	g.Attach(c0, p(e).X)
+	moveMouse(c0.Loc().Min)
 }
 
 func (g *Grid) dragTag(c *Col, t *tag.Tag, e mouse.Event, mousein <-chan mouse.Event) {
-	detachtag(g, c, t)
+	c.detach(c.ID(t))
 	for e = range mousein {
-		e = readmouse(e)
+		e = readmouse0(e)
 		if down == 0 {
 			break
 		}
 	}
-	e.X += float32(act.Sp.X)
-	e.Y += float32(act.Sp.Y)
-	pt.X = int(e.X)
-	pt.Y = int(e.Y)
 	activate(p(e), g)
-	xx.sweep = false
-	xx.srcCol.fill()
-	if xx.src == nil {
+	c.fill()
+	if t == nil {
 		return
 	}
-	c = actCol
-	c.Attach(xx.src, pt.Y)
-	moveMouse(xx.src.Loc().Min)
-}
-
-func detachtag(g *Grid, c *Col, t *tag.Tag) {
-	xx.srcCol = c
-	xx.src = t
-	c.detach(c.ID(t))
-}
-func detachcol(g *Grid, c *Col) {
-	xx.sweepCol = true
-	xx.srcCol = c
-	xx.sweepCol = true
-	xx.src = nil
-	g.detach(g.ID(c))
-	g.fill()
-}
-
-func detacxh(g *Grid, it interface{}) {
-	switch item := it.(type) {
-	case nil:
-		g.aerr("cant detach nil")
-	case *Col:
-		xx.srcCol = item
-		xx.src = nil
-		xx.sweepCol = true
-		g.detach(g.ID(item))
-		g.fill()
-	case *tag.Tag:
-		g.aerr("%#v\n", item)
-		panic("constraint violation")
-	case *win.Win:
-		xx.srcCol = actCol
-		xx.src = actTag
-		xx.srcCol.detach(xx.srcCol.ID(xx.src))
-	case interface{}:
-		g.aerr("%#v\n", item)
-		panic("constraint violation")
-	}
-}
-
-type Cursor struct {
-	sweep    bool
-	sweepCol bool
-	srcCol   *Col
-	src      Plane
+	actCol.Attach(t, p(e).Y)
+	moveMouse(t.Loc().Min)
 }
