+++ mouse_other.go
@@ -1,0 +1,14 @@
+// +build !linux
+
+package main
+
+import (
+	"image"
+
+	"github.com/as/cursor"
+	"github.com/as/ms/win"
+)
+
+func moveMouse(pt image.Point) {
+	cursor.MoveTo(win.ClientAbs().Min.Add(pt))
+}
